Struct azure_svc_blobstorage::package_2021_12::blob::query::Headers

pub struct Headers<'a>(/* private fields */);

Implementations

impl<'a> Headers<'a>

pub fn last_modified(&self) -> Result<OffsetDateTime>

Returns the date and time the container was last modified. Any operation that modifies the blob, including an update of the blob’s metadata or properties, changes the last-modified time of the blob.

pub fn x_ms_meta(&self) -> Result<&str>

pub fn content_length(&self) -> Result<i64>

The number of bytes present in the response body.

pub fn content_type(&self) -> Result<&str>

The media type of the body of the response. For Download Blob this is ‘application/octet-stream’

pub fn content_range(&self) -> Result<&str>

Indicates the range of bytes returned in the event that the client requested a subset of the blob by setting the ‘Range’ request header.

pub fn e_tag(&self) -> Result<&str>

The ETag contains a value that you can use to perform operations conditionally. If the request version is 2011-08-18 or newer, the ETag value will be in quotes.

pub fn content_md5(&self) -> Result<&str>

If the blob has an MD5 hash and this operation is to read the full blob, this response header is returned so that the client can check for message content integrity.

pub fn content_encoding(&self) -> Result<&str>

This header returns the value that was specified for the Content-Encoding request header

pub fn cache_control(&self) -> Result<&str>

This header is returned if it was previously specified for the blob.

pub fn content_disposition(&self) -> Result<&str>

This header returns the value that was specified for the ‘x-ms-blob-content-disposition’ header. The Content-Disposition response header field conveys additional information about how to process the response payload, and also can be used to attach additional metadata. For example, if set to attachment, it indicates that the user-agent should not display the response, but instead show a Save As dialog with a filename other than the blob name specified.

pub fn content_language(&self) -> Result<&str>

This header returns the value that was specified for the Content-Language request header.

pub fn x_ms_blob_sequence_number(&self) -> Result<i64>

The current sequence number for a page blob. This header is not returned for block blobs or append blobs

pub fn x_ms_blob_type(&self) -> Result<&str>

The blob’s type.

pub fn x_ms_copy_completion_time(&self) -> Result<OffsetDateTime>

Conclusion time of the last attempted Copy Blob operation where this blob was the destination blob. This value can specify the time of a completed, aborted, or failed copy attempt. This header does not appear if a copy is pending, if this blob has never been the destination in a Copy Blob operation, or if this blob has been modified after a concluded Copy Blob operation using Set Blob Properties, Put Blob, or Put Block List.

pub fn x_ms_copy_status_description(&self) -> Result<&str>

Only appears when x-ms-copy-status is failed or pending. Describes the cause of the last fatal or non-fatal copy operation failure. This header does not appear if this blob has never been the destination in a Copy Blob operation, or if this blob has been modified after a concluded Copy Blob operation using Set Blob Properties, Put Blob, or Put Block List

pub fn x_ms_copy_id(&self) -> Result<&str>

String identifier for this copy operation. Use with Get Blob Properties to check the status of this copy operation, or pass to Abort Copy Blob to abort a pending copy.

pub fn x_ms_copy_progress(&self) -> Result<&str>

Contains the number of bytes copied and the total bytes in the source in the last attempted Copy Blob operation where this blob was the destination blob. Can show between 0 and Content-Length bytes copied. This header does not appear if this blob has never been the destination in a Copy Blob operation, or if this blob has been modified after a concluded Copy Blob operation using Set Blob Properties, Put Blob, or Put Block List

pub fn x_ms_copy_source(&self) -> Result<&str>

URL up to 2 KB in length that specifies the source blob or file used in the last attempted Copy Blob operation where this blob was the destination blob. This header does not appear if this blob has never been the destination in a Copy Blob operation, or if this blob has been modified after a concluded Copy Blob operation using Set Blob Properties, Put Blob, or Put Block List.

pub fn x_ms_copy_status(&self) -> Result<&str>

State of the copy operation identified by x-ms-copy-id.

pub fn x_ms_lease_duration(&self) -> Result<&str>

When a blob is leased, specifies whether the lease is of infinite or fixed duration.

pub fn x_ms_lease_state(&self) -> Result<&str>

Lease state of the blob.

pub fn x_ms_lease_status(&self) -> Result<&str>

The current lease status of the blob.

pub fn x_ms_client_request_id(&self) -> Result<&str>

If a client request id header is sent in the request, this header will be present in the response with the same value.

pub fn x_ms_request_id(&self) -> Result<&str>

This header uniquely identifies the request that was made and can be used for troubleshooting the request.

pub fn x_ms_version(&self) -> Result<&str>

Indicates the version of the Blob service used to execute the request. This header is returned for requests made against version 2009-09-19 and above.

pub fn accept_ranges(&self) -> Result<&str>

Indicates that the service supports requests for partial blob content.

pub fn date(&self) -> Result<OffsetDateTime>

UTC date/time value generated by the service that indicates the time at which the response was initiated

pub fn x_ms_blob_committed_block_count(&self) -> Result<i32>

The number of committed blocks present in the blob. This header is returned only for append blobs.

pub fn x_ms_server_encrypted(&self) -> Result<bool>

The value of this header is set to true if the blob data and application metadata are completely encrypted using the specified algorithm. Otherwise, the value is set to false (when the blob is unencrypted, or if only parts of the blob/application metadata are encrypted).

pub fn x_ms_encryption_key_sha256(&self) -> Result<&str>

The SHA-256 hash of the encryption key used to encrypt the blob. This header is only returned when the blob was encrypted with a customer-provided key.

pub fn x_ms_encryption_scope(&self) -> Result<&str>

Returns the name of the encryption scope used to encrypt the blob contents and application metadata. Note that the absence of this header implies use of the default account encryption scope.

pub fn x_ms_blob_content_md5(&self) -> Result<&str>

If the blob has a MD5 hash, and if request contains range header (Range or x-ms-range), this response header is returned with the value of the whole blob’s MD5 value. This value may or may not be equal to the value returned in Content-MD5 header, with the latter calculated from the requested range

pub fn x_ms_content_crc64(&self) -> Result<&str>

If the request is to read a specified range and the x-ms-range-get-content-crc64 is set to true, then the request returns a crc64 for the range, as long as the range size is less than or equal to 4 MB. If both x-ms-range-get-content-crc64 and x-ms-range-get-content-md5 is specified in the same request, it will fail with 400(Bad Request)