Struct azure_svc_blobstorage::package_2021_12::block_blob::commit_block_list::RequestBuilder

pub struct RequestBuilder { /* private fields */ }

RequestBuilder provides a mechanism for setting optional parameters on a request.

Each RequestBuilder parameter method call returns Self, so setting of multiple parameters can be chained.

To finalize and submit the request, invoke .await, which which will convert the RequestBuilder into a future executes the request and returns a Result with the parsed response.

In order to execute the request without polling the service until the operation completes, use .send().await instead.

If you need lower-level access to the raw response details (e.g. to inspect response headers or raw body data) then you can finalize the request using the RequestBuilder::send() method which returns a future that resolves to a lower-level Response value.

Implementations

impl RequestBuilder

pub fn timeout(self, timeout: i64) -> Self

The timeout parameter is expressed in seconds. For more information, see Setting Timeouts for Blob Service Operations.

pub fn x_ms_blob_cache_control( self, x_ms_blob_cache_control: impl Into<String>, ) -> Self

Optional. Sets the blob’s cache control. If specified, this property is stored with the blob and returned with a read request.

pub fn x_ms_blob_content_type( self, x_ms_blob_content_type: impl Into<String>, ) -> Self

Optional. Sets the blob’s content type. If specified, this property is stored with the blob and returned with a read request.

pub fn x_ms_blob_content_encoding( self, x_ms_blob_content_encoding: impl Into<String>, ) -> Self

Optional. Sets the blob’s content encoding. If specified, this property is stored with the blob and returned with a read request.

pub fn x_ms_blob_content_language( self, x_ms_blob_content_language: impl Into<String>, ) -> Self

Optional. Set the blob’s content language. If specified, this property is stored with the blob and returned with a read request.

pub fn x_ms_blob_content_md5( self, x_ms_blob_content_md5: impl Into<String>, ) -> Self

Optional. An MD5 hash of the blob content. Note that this hash is not validated, as the hashes for the individual blocks were validated when each was uploaded.

pub fn content_md5(self, content_md5: impl Into<String>) -> Self

Specify the transactional md5 for the body, to be validated by the service.

pub fn x_ms_content_crc64(self, x_ms_content_crc64: impl Into<String>) -> Self

Specify the transactional crc64 for the body, to be validated by the service.

pub fn x_ms_meta(self, x_ms_meta: impl Into<String>) -> Self

Optional. Specifies a user-defined name-value pair associated with the blob. If no name-value pairs are specified, the operation will copy the metadata from the source blob or file to the destination blob. If one or more name-value pairs are specified, the destination blob is created with the specified metadata, and metadata is not copied from the source blob or file. Note that beginning with version 2009-09-19, metadata names must adhere to the naming rules for C# identifiers. See Naming and Referencing Containers, Blobs, and Metadata for more information.

pub fn x_ms_lease_id(self, x_ms_lease_id: impl Into<String>) -> Self

If specified, the operation only succeeds if the resource’s lease is active and matches this ID.

pub fn x_ms_blob_content_disposition( self, x_ms_blob_content_disposition: impl Into<String>, ) -> Self

Optional. Sets the blob’s Content-Disposition header.

pub fn x_ms_encryption_key(self, x_ms_encryption_key: impl Into<String>) -> Self

Optional. Specifies the encryption key to use to encrypt the data provided in the request. If not specified, encryption is performed with the root account encryption key. For more information, see Encryption at Rest for Azure Storage Services.

pub fn x_ms_encryption_key_sha256( self, x_ms_encryption_key_sha256: impl Into<String>, ) -> Self

The SHA-256 hash of the provided encryption key. Must be provided if the x-ms-encryption-key header is provided.

pub fn x_ms_encryption_algorithm( self, x_ms_encryption_algorithm: impl Into<String>, ) -> Self

The algorithm used to produce the encryption key hash. Currently, the only accepted value is “AES256”. Must be provided if the x-ms-encryption-key header is provided.

pub fn x_ms_encryption_scope( self, x_ms_encryption_scope: impl Into<String>, ) -> Self

Optional. Version 2019-07-07 and later. Specifies the name of the encryption scope to use to encrypt the data provided in the request. If not specified, encryption is performed with the default account encryption scope. For more information, see Encryption at Rest for Azure Storage Services.

pub fn x_ms_access_tier(self, x_ms_access_tier: impl Into<String>) -> Self

Optional. Indicates the tier to be set on the blob.

pub fn if_modified_since( self, if_modified_since: impl Into<OffsetDateTime>, ) -> Self

Specify this header value to operate only on a blob if it has been modified since the specified date/time.

pub fn if_unmodified_since( self, if_unmodified_since: impl Into<OffsetDateTime>, ) -> Self

Specify this header value to operate only on a blob if it has not been modified since the specified date/time.

pub fn if_match(self, if_match: impl Into<String>) -> Self

Specify an ETag value to operate only on blobs with a matching value.

pub fn if_none_match(self, if_none_match: impl Into<String>) -> Self

Specify an ETag value to operate only on blobs without a matching value.

pub fn x_ms_if_tags(self, x_ms_if_tags: impl Into<String>) -> Self

Specify a SQL where clause on blob tags to operate only on blobs with a matching value.

pub fn x_ms_client_request_id( self, x_ms_client_request_id: impl Into<String>, ) -> Self

Provides a client-generated, opaque value with a 1 KB character limit that is recorded in the analytics logs when storage analytics logging is enabled.

pub fn x_ms_tags(self, x_ms_tags: impl Into<String>) -> Self

Optional. Used to set blob tags in various blob operations.

pub fn x_ms_immutability_policy_until_date( self, x_ms_immutability_policy_until_date: impl Into<OffsetDateTime>, ) -> Self

Specifies the date time when the blobs immutability policy is set to expire.

pub fn x_ms_immutability_policy_mode( self, x_ms_immutability_policy_mode: impl Into<String>, ) -> Self

Specifies the immutability policy mode to set on the blob.

pub fn x_ms_legal_hold(self, x_ms_legal_hold: bool) -> Self

Specified if a legal hold should be set on the blob.

pub fn send(self) -> BoxFuture<'static, Result<Response>>

Returns a future that sends the request and returns a Response object that provides low-level access to full response details.

You should typically use .await (which implicitly calls IntoFuture::into_future()) to finalize and send requests rather than send(). However, this function can provide more flexibility when required.