Struct azure_svc_blobstorage::package_2021_12::block_blob::Client

pub struct Client(/* private fields */);

Implementations

impl Client

pub fn upload( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_blob_type: impl Into<String>, body: impl Into<Value>, content_length: i64, ) -> RequestBuilder

The Upload Block Blob operation updates the content of an existing block blob. Updating an existing block blob overwrites any existing metadata on the blob. Partial updates are not supported with Put Blob; the content of the existing blob is overwritten with the content of the new blob. To perform a partial update of the content of a block blob, use the Put Block List operation.

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_blob_type: Specifies the type of blob to create: block blob, page blob, or append blob.

body: Initial data

content_length: The length of the request.

pub fn put_blob_from_url( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_blob_type: impl Into<String>, content_length: i64, x_ms_copy_source: impl Into<String>, ) -> RequestBuilder

The Put Blob from URL operation creates a new Block Blob where the contents of the blob are read from a given URL. This API is supported beginning with the 2020-04-08 version. Partial updates are not supported with Put Blob from URL; the content of an existing blob is overwritten with the content of the new blob. To perform partial updates to a block blob’s contents using a source URL, use the Put Block from URL API in conjunction with Put Block List.

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_blob_type: Specifies the type of blob to create: block blob, page blob, or append blob.

content_length: The length of the request.

x_ms_copy_source: Specifies the name of the source page blob snapshot. This value is a URL of up to 2 KB in length that specifies a page blob snapshot. The value should be URL-encoded as it would appear in a request URI. The source blob must either be public or must be authenticated via a shared access signature.

pub fn stage_block( &self, container_name: impl Into<String>, blob: impl Into<String>, blockid: impl Into<String>, content_length: i64, body: impl Into<Value>, ) -> RequestBuilder

The Stage Block operation creates a new block to be committed as part of a blob

Arguments:

container_name: The container name.

blob: The blob name.

blockid: A valid Base64 string value that identifies the block. Prior to encoding, the string must be less than or equal to 64 bytes in size. For a given blob, the length of the value specified for the blockid parameter must be the same size for each block.

content_length: The length of the request.

body: Initial data

pub fn stage_block_from_url( &self, container_name: impl Into<String>, blob: impl Into<String>, blockid: impl Into<String>, content_length: i64, x_ms_copy_source: impl Into<String>, ) -> RequestBuilder

The Stage Block operation creates a new block to be committed as part of a blob where the contents are read from a URL.

Arguments:

container_name: The container name.

blob: The blob name.

blockid: A valid Base64 string value that identifies the block. Prior to encoding, the string must be less than or equal to 64 bytes in size. For a given blob, the length of the value specified for the blockid parameter must be the same size for each block.

content_length: The length of the request.

x_ms_copy_source: Specify a URL to the copy source.

pub fn get_block_list( &self, container_name: impl Into<String>, blob: impl Into<String>, blocklisttype: impl Into<String>, ) -> RequestBuilder

The Get Block List operation retrieves the list of blocks that have been uploaded as part of a block blob

Arguments:

container_name: The container name.

blob: The blob name.

blocklisttype: Specifies whether to return the list of committed blocks, the list of uncommitted blocks, or both lists together.

pub fn commit_block_list( &self, container_name: impl Into<String>, blob: impl Into<String>, blocks: impl Into<BlockLookupList>, ) -> RequestBuilder

The Commit Block List operation writes a blob by specifying the list of block IDs that make up the blob. In order to be written as part of a blob, a block must have been successfully written to the server in a prior Put Block operation. You can call Put Block List to update a blob by uploading only those blocks that have changed, then committing the new and existing blocks together. You can do this by specifying whether to commit a block from the committed block list or from the uncommitted block list, or to commit the most recently uploaded version of the block, whichever list it may belong to.

Arguments:

container_name: The container name.

blob: The blob name.

blocks: Blob Blocks.