Struct azure_svc_blobstorage::package_2021_12::container::Client

pub struct Client(/* private fields */);

Implementations

impl Client

pub fn get_properties( &self, container_name: impl Into<String>, ) -> RequestBuilder

returns all user-defined metadata and system properties for the specified container. The data returned does not include the container’s list of blobs

Arguments:

container_name: The container name.

pub fn create(&self, container_name: impl Into<String>) -> RequestBuilder

creates a new container under the specified account. If the container with the same name already exists, the operation fails

Arguments:

container_name: The container name.

pub fn delete(&self, container_name: impl Into<String>) -> RequestBuilder

operation marks the specified container for deletion. The container and any blobs contained within it are later deleted during garbage collection

Arguments:

container_name: The container name.

pub fn set_metadata(&self, container_name: impl Into<String>) -> RequestBuilder

operation sets one or more user-defined name-value pairs for the specified container.

Arguments:

container_name: The container name.

pub fn get_access_policy( &self, container_name: impl Into<String>, ) -> RequestBuilder

gets the permissions for the specified container. The permissions indicate whether container data may be accessed publicly.

Arguments:

container_name: The container name.

pub fn set_access_policy( &self, container_name: impl Into<String>, ) -> RequestBuilder

sets the permissions for the specified container. The permissions indicate whether blobs in a container may be accessed publicly.

Arguments:

container_name: The container name.

pub fn restore(&self, container_name: impl Into<String>) -> RequestBuilder

Restores a previously-deleted container.

Arguments:

container_name: The container name.

pub fn rename( &self, container_name: impl Into<String>, x_ms_source_container_name: impl Into<String>, ) -> RequestBuilder

Renames an existing container.

Arguments:

container_name: The container name.

x_ms_source_container_name: Required. Specifies the name of the container to rename.

pub fn submit_batch( &self, container_name: impl Into<String>, body: impl Into<Value>, content_length: i64, content_type: impl Into<String>, ) -> RequestBuilder

The Batch operation allows multiple API calls to be embedded into a single HTTP request.

Arguments:

container_name: The container name.

body: Initial data

content_length: The length of the request.

content_type: Required. The value of this header must be multipart/mixed with a batch boundary. Example header value: multipart/mixed; boundary=batch_

pub fn filter_blobs(&self, container_name: impl Into<String>) -> RequestBuilder

The Filter Blobs operation enables callers to list blobs in a container whose tags match a given search expression. Filter blobs searches within the given container.

Arguments:

container_name: The container name.

pub fn acquire_lease( &self, container_name: impl Into<String>, x_ms_lease_action: impl Into<String>, ) -> RequestBuilder

[Update] establishes and manages a lock on a container for delete operations. The lock duration can be 15 to 60 seconds, or can be infinite

Arguments:

container_name: The container name.

x_ms_lease_action: Describes what lease action to take.

pub fn release_lease( &self, container_name: impl Into<String>, x_ms_lease_action: impl Into<String>, x_ms_lease_id: impl Into<String>, ) -> RequestBuilder

[Update] establishes and manages a lock on a container for delete operations. The lock duration can be 15 to 60 seconds, or can be infinite

Arguments:

container_name: The container name.

x_ms_lease_action: Describes what lease action to take.

x_ms_lease_id: Specifies the current lease ID on the resource.

pub fn renew_lease( &self, container_name: impl Into<String>, x_ms_lease_action: impl Into<String>, x_ms_lease_id: impl Into<String>, ) -> RequestBuilder

[Update] establishes and manages a lock on a container for delete operations. The lock duration can be 15 to 60 seconds, or can be infinite

Arguments:

container_name: The container name.

x_ms_lease_action: Describes what lease action to take.

x_ms_lease_id: Specifies the current lease ID on the resource.

pub fn break_lease( &self, container_name: impl Into<String>, x_ms_lease_action: impl Into<String>, ) -> RequestBuilder

[Update] establishes and manages a lock on a container for delete operations. The lock duration can be 15 to 60 seconds, or can be infinite

Arguments:

container_name: The container name.

x_ms_lease_action: Describes what lease action to take.

pub fn change_lease( &self, container_name: impl Into<String>, x_ms_lease_action: impl Into<String>, x_ms_lease_id: impl Into<String>, x_ms_proposed_lease_id: impl Into<String>, ) -> RequestBuilder

[Update] establishes and manages a lock on a container for delete operations. The lock duration can be 15 to 60 seconds, or can be infinite

Arguments:

container_name: The container name.

x_ms_lease_action: Describes what lease action to take.

x_ms_lease_id: Specifies the current lease ID on the resource.

x_ms_proposed_lease_id: Proposed lease ID, in a GUID string format. The Blob service returns 400 (Invalid request) if the proposed lease ID is not in the correct format. See Guid Constructor (String) for a list of valid GUID string formats.

pub fn list_blob_flat_segment( &self, container_name: impl Into<String>, ) -> RequestBuilder

[Update] The List Blobs operation returns a list of the blobs under the specified container

Arguments:

container_name: The container name.

pub fn list_blob_hierarchy_segment( &self, container_name: impl Into<String>, delimiter: impl Into<String>, ) -> RequestBuilder

[Update] The List Blobs operation returns a list of the blobs under the specified container

Arguments:

container_name: The container name.

delimiter: When the request includes this parameter, the operation returns a BlobPrefix element in the response body that acts as a placeholder for all blobs whose names begin with the same substring up to the appearance of the delimiter character. The delimiter may be a single character or a string.

pub fn get_account_info( &self, container_name: impl Into<String>, ) -> RequestBuilder

Returns the sku name and account kind

Arguments:

container_name: The container name.