Struct azure_svc_blobstorage::package_2021_12::page_blob::Client

pub struct Client(/* private fields */);

Implementations

impl Client

pub fn create( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_blob_type: impl Into<String>, content_length: i64, x_ms_blob_content_length: i64, ) -> RequestBuilder

The Create operation creates a new page blob.

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_blob_type: Specifies the type of blob to create: block blob, page blob, or append blob.

content_length: The length of the request.

x_ms_blob_content_length: This header specifies the maximum size for the page blob, up to 1 TB. The page blob size must be aligned to a 512-byte boundary.

pub fn upload_pages( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_page_write: impl Into<String>, body: impl Into<Value>, content_length: i64, ) -> RequestBuilder

The Upload Pages operation writes a range of pages to a page blob

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_page_write: Required. You may specify one of the following options:

Update: Writes the bytes specified by the request body into the specified range. The Range and Content-Length headers must match to perform the update.

Clear: Clears the specified range and releases the space used in storage for that range. To clear a range, set the Content-Length header to zero, and the Range header to a value that indicates the range to clear, up to maximum blob size.

body: Initial data

content_length: The length of the request.

pub fn clear_pages( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_page_write: impl Into<String>, content_length: i64, ) -> RequestBuilder

The Clear Pages operation clears a set of pages from a page blob

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_page_write: Required. You may specify one of the following options:

Update: Writes the bytes specified by the request body into the specified range. The Range and Content-Length headers must match to perform the update.

Clear: Clears the specified range and releases the space used in storage for that range. To clear a range, set the Content-Length header to zero, and the Range header to a value that indicates the range to clear, up to maximum blob size.

content_length: The length of the request.

pub fn upload_pages_from_url( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_page_write: impl Into<String>, x_ms_copy_source: impl Into<String>, x_ms_source_range: impl Into<String>, content_length: i64, x_ms_range: impl Into<String>, ) -> RequestBuilder

The Upload Pages operation writes a range of pages to a page blob where the contents are read from a URL

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_page_write: Required. You may specify one of the following options:

Update: Writes the bytes specified by the request body into the specified range. The Range and Content-Length headers must match to perform the update.

Clear: Clears the specified range and releases the space used in storage for that range. To clear a range, set the Content-Length header to zero, and the Range header to a value that indicates the range to clear, up to maximum blob size.

x_ms_copy_source: Specify a URL to the copy source.

x_ms_source_range: Bytes of source data in the specified range. The length of this range should match the ContentLength header and x-ms-range/Range destination range header.

content_length: The length of the request.

x_ms_range: The range of bytes to which the source range would be written. The range should be 512 aligned and range-end is required.

pub fn get_page_ranges( &self, container_name: impl Into<String>, blob: impl Into<String>, ) -> RequestBuilder

The Get Page Ranges operation returns the list of valid page ranges for a page blob or snapshot of a page blob

Arguments:

container_name: The container name.

blob: The blob name.

pub fn get_page_ranges_diff( &self, container_name: impl Into<String>, blob: impl Into<String>, ) -> RequestBuilder

The Get Page Ranges Diff operation returns the list of valid page ranges for a page blob that were changed between target blob and previous snapshot.

Arguments:

container_name: The container name.

blob: The blob name.

pub fn resize( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_blob_content_length: i64, ) -> RequestBuilder

Resize the Blob

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_blob_content_length: This header specifies the maximum size for the page blob, up to 1 TB. The page blob size must be aligned to a 512-byte boundary.

pub fn update_sequence_number( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_sequence_number_action: impl Into<String>, ) -> RequestBuilder

Update the sequence number of the blob

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_sequence_number_action: Required if the x-ms-blob-sequence-number header is set for the request. This property applies to page blobs only. This property indicates how the service should modify the blob’s sequence number

pub fn copy_incremental( &self, container_name: impl Into<String>, blob: impl Into<String>, x_ms_copy_source: impl Into<String>, ) -> RequestBuilder

The Copy Incremental operation copies a snapshot of the source page blob to a destination page blob. The snapshot is copied such that only the differential changes between the previously copied snapshot are transferred to the destination. The copied snapshots are complete copies of the original snapshot and can be read or copied from as usual. This API is supported since REST version 2016-05-31.

Arguments:

container_name: The container name.

blob: The blob name.

x_ms_copy_source: Specifies the name of the source page blob snapshot. This value is a URL of up to 2 KB in length that specifies a page blob snapshot. The value should be URL-encoded as it would appear in a request URI. The source blob must either be public or must be authenticated via a shared access signature.