Module azure_svc_blobstorage::package_2021_12::models

Modules

arrow_configuration

blob_properties_internal

blob_tags

block_list

filter_blob_segment

geo_replication

list_containers_segment_response

query_request

storage_service_properties

Structs

AccessPolicy

An Access policy

ArrowConfiguration

Groups the settings used for formatting the response if the response should be Arrow formatted.

ArrowField

Groups settings regarding specific field of an arrow schema

BlobFlatListSegment

BlobHierarchyListSegment

BlobItemInternal

An Azure Storage blob

BlobMetadata

BlobName

BlobPrefix

BlobPropertiesInternal

Properties of a blob

BlobTag

BlobTags

Blob tags

Block

Represents a single block in a block blob. It describes the block’s ID and size.

BlockList

BlockLookupList

ClearRange

ContainerItem

An Azure Storage container

ContainerMetadata

ContainerProperties

Properties of a container

CorsRule

CORS is an HTTP feature that enables a web application running under one domain to access resources in another domain. Web browsers implement a security restriction known as same-origin policy that prevents a web page from calling APIs in a different domain; CORS provides a secure way to allow one domain (the origin domain) to call APIs in another domain

DelimitedTextConfiguration

Groups the settings used for interpreting the blob data if the blob is delimited text formatted.

FilterBlobItem

Blob info from a Filter Blobs API call

FilterBlobSegment

The result of a Filter Blobs API call

GeoReplication

Geo-Replication information for the Secondary Storage Service

JsonTextConfiguration

json text configuration

KeyInfo

Key information

ListBlobsFlatSegmentResponse

An enumeration of blobs

ListBlobsHierarchySegmentResponse

An enumeration of blobs

ListContainersSegmentResponse

An enumeration of containers

Logging

Azure Analytics Logging settings.

Metrics

a summary of request statistics grouped by API in hour or minute aggregates for blobs

ObjectReplicationMetadata

PageList

the list of pages

PageRange

ParquetConfiguration

parquet configuration

QueryFormat

QueryRequest

Groups the set of query request settings.

QuerySerialization

RetentionPolicy

the retention policy which determines how long the associated data should persist

SignedIdentifier

signed identifier

StaticWebsite

The properties that enable an account to host a static website

StorageError

StorageServiceProperties

Storage Service Properties.

StorageServiceStats

Stats for the storage service.

UserDelegationKey

A user delegation key

Enums

AccessTier

ArchiveStatus

CopyStatus

ErrorCode

Error codes returned by the service

LeaseDuration

LeaseState

LeaseStatus

PublicAccessType

QueryType

The quick query format type.

RehydratePriority

If an object is in rehydrate pending state then this header is returned with priority of rehydrate. Valid values are High and Standard.

Type Aliases

SignedIdentifiers